+++ src/Bridge/Symfony/ElasticallyBundle.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the jolicode/elastically library.
+ *
+ * (c) JoliCode <coucou@jolicode.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace JoliCode\Elastically\Bridge\Symfony;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ElasticallyBundle extends Bundle
+{
+}