+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Drop support support for Symfony 4.*
 - Drop support for PHP 7.4
+- Add a bundle to ease integration with Symfony
 
 ## [1.5.4] - 2022-12-16